Remove template files from the guide
When the guide was initially created, it seems that the template
files were left in the project. The guide only uses the code
inside of the reading and bookstore folders. The other files
were leftover from the initial creation when a template was copied.
This commit removes those files as they are likely to confuse the
reader and have nothing to do with the guide.