Simplify test cases for clean compile
The existing test cases in this guide were not run in the CI build
and were out of date. Because the tests were never run in the CI system,
it is difficult to know when they broke or if they ever worked.
Instead of rewriting the tests, this commit opts to simplify the
tests back to a base state that would come from Spring Initializr.

The reasoning for the simplification is that the README and guide
instructions do not mention the testing code at all. Instead, there
is an additional guide published on testing the web layer that the
user could refer to. The getting started guides aim to simplify as
much as possible and not include code that is not instructed on.
This is a pattern followed throughout multiple guides.