Update dependency versions for bookstore project
This commit updates all dependencies for the bookstore project.
This includes both maven and gralde builds in both initial
and complete folders. This manual upgrade method is favored
over dependabot at this time because multiple updates should be
done simultaneously. Going forward this requirement should
be handled with a more appropriate dependabot config file.

The dependency for springCloudVersion was changed from set to =
so that dependabot can recognize this dependency for future updates.